docker/platform-management/hackathon/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.4.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
redis==5.0.1
prometheus-client==0.19.0
pytest==7.4.3
httpx==0.25.2
